common/pyproject.toml
[tool.poetry]
name = "common"
version = "0.1.0"
description = ""
authors = ["Lawrence Nevvon"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
ptw = "^1.0.1"
Flask = "^2.2.2"
Flask-RESTful = "^0.3.9"
requests = "^2.28.1"
pytest-flask = "^1.2.0"
isort = "^5.10.1"
mysql-connector-python = "^8.0.31"
freezegun = "^1.3.1"
paramiko = "^3.4.0"
pytest = "^8.3.5"

[tool.poetry.dev-dependencies]
autopep8 = "^2.0.0"
pytest-cov = "^4.0.0"
pytest-watch = "^4.2.0"


[tool.poetry.group.dev.dependencies]
pytest-env = "^1.1.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest_env]
DB_CONFIG = "{'DBHOST': '127.0.0.1', 'DBUSER': 'root','DBPASSWORD':'password','DBNAME':'integrations_db'}"
